feature-erms-2248
as of ERMS-2248, export of custom and private properties have been generalised and introduced for currentMembers and currentSubscriptions view
