+++ app/CHANGELOG-TMP.md
@@ -2,6 +2,8 @@
 
 **Ticket    Date    Branch      Author  Feature/Bug     Description/Keywords**
 
+2248    13.03.2020  hbz-dev     Andreas Feature     Generische Ausgabe allgemeiner und privater Merkmale angelegt und für Lizenzen und Konsortenliste eingerichtet
+
 2186    12.03.2020  hbz-dev     Christian   Feature Zugangskonfiguration: Löschen grundsätzlich unterbinden, wenn Verknüpfungen vorliegen
     
 2203    12.03.2020  rc1.2       Moe     Bug         pick&choose nicht möglich
