+++ app/application.properties
@@ -1,7 +1,7 @@
 #Grails Metadata file
-#Wed Mar 11 14:37:02 CET 2020
-app.buildDate=11.03.2020 14\:37
-app.buildNumber=7819
+#Fri Mar 13 11:06:41 CET 2020
+app.buildDate=13.03.2020 11\:06
+app.buildNumber=7820
 app.buildProfile=development
 app.grails.version=2.5.6
 app.name=laser
