+++ app/grails-app/services/com/k_int/kbplus/OrganisationService.groovy
@@ -86,11 +86,7 @@ class OrganisationService {
         RefdataValue responsibleAdmin = RefdataValue.getByValueAndCategory('Responsible Admin', RDConstants.PERSON_FUNCTION)
         RefdataValue billingContact = RefdataValue.getByValueAndCategory('Functional Contact Billing Adress', RDConstants.PERSON_FUNCTION)
         titles.addAll(['ISIL','WIB-ID','EZB-ID',generalContact.getI10n('value'),responsibleAdmin.getI10n('value'),billingContact.getI10n('value')])
-        def propList = PropertyDefinition.findAllPublicAndPrivateOrgProp(contextService.getOrg())
-        propList.sort { a, b -> a.name.compareToIgnoreCase b.name}
-        propList.each {
-            titles.add(it.name)
-        }
+        titles.addAll(exportService.loadPropListHeaders(PropertyDefinition.ORG_PROP,contextService.getOrg()))
         List orgData = []
         Map<Org,Map<String,String>> identifiers = [:]
         List identifierList = Identifier.executeQuery("select ident, ident.org from Identifier ident where ident.org in (:orgs) and ident.ns.ns in (:namespaces)",[orgs:orgs,namespaces:['wibid','ezb','ISIL']])
@@ -159,48 +155,7 @@ class OrganisationService {
                     row.add([field: furtherData.responsibleAdmin ?: '', style: null])
                     //Billing contact
                     row.add([field: furtherData.billingContact ?: '', style: null])
-                    propList.each { pd ->
-                        def value = ''
-                        org.customProperties.each{ prop ->
-                            if(prop.type.descr == pd.descr && prop.type == pd) {
-                                if(prop.type.type == Integer.toString()){
-                                    value = prop.intValue.toString()
-                                }
-                                else if (prop.type.type == String.toString()){
-                                    value = prop.stringValue ?: ''
-                                }
-                                else if (prop.type.type == BigDecimal.toString()){
-                                    value = prop.decValue.toString()
-                                }
-                                else if (prop.type.type == Date.toString()){
-                                    value = prop.dateValue.toString()
-                                }
-                                else if (prop.type.type == RefdataValue.toString()) {
-                                    value = prop.refValue?.getI10n('value') ?: ''
-                                }
-                            }
-                        }
-                        org.privateProperties.each{ prop ->
-                            if(prop.type.descr == pd.descr && prop.type == pd) {
-                                if(prop.type.type == Integer.toString()){
-                                    value = prop.intValue.toString()
-                                }
-                                else if (prop.type.type == String.toString()){
-                                    value = prop.stringValue ?: ''
-                                }
-                                else if (prop.type.type == BigDecimal.toString()){
-                                    value = prop.decValue.toString()
-                                }
-                                else if (prop.type.type == Date.toString()){
-                                    value = prop.dateValue.toString()
-                                }
-                                else if (prop.type.type == RefdataValue.toString()) {
-                                    value = prop.refValue?.getI10n('value') ?: ''
-                                }
-                            }
-                        }
-                        row.add([field: value, style: null])
-                    }
+                    row.addAll(exportService.processPropertyListValues(PropertyDefinition.ORG_PROP,contextService.org,format,org))
                     orgData.add(row)
                 }
                 Map sheetData = [:]
@@ -243,48 +198,7 @@ class OrganisationService {
                     row.add(furtherData.responsibleAdmin ?: '')
                     //Billing contact
                     row.add(furtherData.billingContact ?: '')
-                    propList.each { pd ->
-                        def value = ''
-                        org.customProperties.each{ prop ->
-                            if(prop.type.descr == pd.descr && prop.type == pd) {
-                                if(prop.type.type == Integer.toString()){
-                                    value = prop.intValue.toString()
-                                }
-                                else if (prop.type.type == String.toString()){
-                                    value = prop.stringValue ?: ''
-                                }
-                                else if (prop.type.type == BigDecimal.toString()){
-                                    value = prop.decValue.toString()
-                                }
-                                else if (prop.type.type == Date.toString()){
-                                    value = prop.dateValue.toString()
-                                }
-                                else if (prop.type.type == RefdataValue.toString()) {
-                                    value = prop.refValue?.getI10n('value') ?: ''
-                                }
-                            }
-                        }
-                        org.privateProperties.each{ prop ->
-                            if(prop.type.descr == pd.descr && prop.type == pd) {
-                                if(prop.type.type == Integer.toString()){
-                                    value = prop.intValue.toString()
-                                }
-                                else if (prop.type.type == String.toString()){
-                                    value = prop.stringValue ?: ''
-                                }
-                                else if (prop.type.type == BigDecimal.toString()){
-                                    value = prop.decValue.toString()
-                                }
-                                else if (prop.type.type == Date.toString()){
-                                    value = prop.dateValue.toString()
-                                }
-                                else if (prop.type.type == RefdataValue.toString()) {
-                                    value = prop.refValue?.getI10n('value') ?: ''
-                                }
-                            }
-                        }
-                        row.add(value.replaceAll(',',';'))
-                    }
+                    row.addAll(exportService.processPropertyListValues(PropertyDefinition.ORG_PROP,contextService.org,format,org))
                     orgData.add(row)
                 }
                 return exportService.generateSeparatorTableString(titles,orgData,',')
